Refactored the receipt page, so the components are now in separate files
